removed scope from controllers

+++ app/controllers/internal_api/v1/timesheet_entry_controller.rb
@@ -3,31 +3,31 @@
 class InternalApi::V1::TimesheetEntryController < InternalApi::V1::ApplicationController
   include Timesheet
 
-  skip_after_action :verify_authorized
-  after_action :verify_policy_scoped
+  skip_after_action :verify_authorized, only: [:index]
+  after_action :verify_policy_scoped, only: [:index]
 
   def index
-    scope = policy_scope(TimesheetEntry)
-    timesheet_entries = scope.where(user_id: params[:user_id] || current_user.id).during(params[:from], params[:to])
+    timesheet_entries = policy_scope(TimesheetEntry)
+    timesheet_entries = timesheet_entries.where(user_id: params[:user_id] || current_user.id).during(params[:from], params[:to])
     entries = formatted_entries_by_date(timesheet_entries)
     render json: { entries: entries }
   end
 
   def create
-    scope = policy_scope(TimesheetEntry)
+    authorize TimesheetEntry
     timesheet_entry = current_project.timesheet_entries.new(timesheet_entry_params)
     timesheet_entry.user = current_user
     render json: { notice: "New entry added for #{timesheet_entry.work_date.to_date}", entry: timesheet_entry.formatted_entry } if timesheet_entry.save
   end
 
   def update
-    scope = policy_scope(current_timesheet_entry)
+    authorize current_timesheet_entry
     current_timesheet_entry.project = current_project
     render json: { notice: "Timesheet updated", entry: current_timesheet_entry.formatted_entry }, status: :ok if current_timesheet_entry.update(timesheet_entry_params)
   end
 
   def destroy
-    scope = policy_scope(current_timesheet_entry)
+    authorize current_timesheet_entry
     render json: { notice: "Timesheet deleted" } if current_timesheet_entry.destroy
   end
